Update user-access.md
When we say " Each user in your organization must have a Teams license before they can use Teams", this is in direct conflict with ability of user to joining meetings (refer https://docs.microsoft.com/en-us/MicrosoftTeams/quick-start-meetings-live-events#get-licenses-for-meetings-and-live-events). Can we either put a exception pointing to the meetings document or modify the language to "Each user in your organization must have a Teams license before they can access Teams".